apps/backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "labordaten-backend"
version = "0.1.0"
description = "Lokales Backend für die Laboranwendung Labordaten"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.13,<2",
  "fastapi>=0.111,<1",
  "pydantic-settings>=2.2,<3",
  "python-multipart>=0.0.9,<1",
  "sqlalchemy>=2.0,<3",
  "uvicorn[standard]>=0.30,<1"
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27,<1",
  "pytest>=8.2,<9"
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
